Refactor task_runner module location
Moved TaskRunner from lib/cli_tools to cli directory.
