+++ foundrytools_cli_2/cli/os_2/cli.py
@@ -13,7 +13,7 @@ from foundrytools_cli_2.cli.os_2.options import (
 )
 from foundrytools_cli_2.cli.shared_callbacks import validate_params
 from foundrytools_cli_2.cli.shared_options import base_options
-from foundrytools_cli_2.lib.cli_tools.task_runner import TaskRunner
+from foundrytools_cli_2.cli.task_runner import TaskRunner
 
 cli = click.Group(help="Utilities for editing the ``OS/2`` table.")
